Support only bootstrapping without running the servers
Also fixing some bugs around when to deploy a staged ddoc. Previously
we only deployed if attached node apps had changed, now we always do it.

In theory we should only deploy a staged ddoc if it's a different ddoc
to what is already there, however until we have a good known piece of
metadata to compare ddocs against there is not a good way to be certain.