Remove `chown_apps` feature
Horticulturalist, IMNSHO, should never be running as root, and
should use `sudo` to escalate privileges temporarily when needed
to start/stop a service.

This changeset removes the `chown_apps` feature, which was only
enabled for `--medic-os` deployments and didn't work unless the
daemon was running as root. It caused errors in unprivileged
deployment scenarios.

Use `sudo` to start/stop packages when running with `--medic-os`.

Toward medic/medic-webapp#4204.