Skipping stages #36
Adds support for stages to be skipped if they've already been run.

This would trigger if you started horti without an action with an
existing deployment document with some already completed stages (we're
somewhat oddly defining a completed stage as one that is written to the
stage log and has other stages after it).

This would not trigger if you started horti with an explicit install,
stage or complete action, as the presumed intention is that you wish to
start a new deployment and overwrite whatever else you had going on.